RAG_Chatbot/requirements.txt
sentence-transformers
qdrant-client
google-generativeai
python-dotenv
Pillow
pyvi
markdownify 
duckduckgo-search 
smolagents 
langchain-community 
faiss-cpu 
langchain 
unstructured 
python-docx 
spaces 
gradio-tools 
langchain-huggingface
flask 
flask-cors 
python-dotenv 
markdownify 
duckduckgo-search 
smolagents 
requests 
beautifulsoup4